Added an IStrategicTheme interface and moved the StrategicThemeState enum to the common domain project.
