+++ Moda.Common/src/Moda.Common.Domain/Interfaces/StrategicManagement/IStrategicTheme.cs
@@ -1,0 +1,11 @@
+﻿using Moda.Common.Domain.Enums.StrategicManagement;
+
+namespace Moda.Common.Domain.Interfaces.StrategicManagement;
+public interface IStrategicTheme
+{
+    Guid Id { get; }
+    int Key { get; }
+    string Name { get; }
+    string Description { get; }
+    StrategicThemeState State { get; }
+}
